+++ lib/core/controlplane.rb
@@ -438,6 +438,18 @@ class Controlplane # rubocop:disable Metrics/ClassLength
     api.reveal_secret(org: org, secret: secret)
   end
 
+  def create_sensitive_secret(secret, data)
+    api.create_sensitive_secret(org: org, secret: secret, data: data)
+  end
+
+  def replace_sensitive_secret_data(secret, data)
+    api.replace_sensitive_secret_data(org: org, secret: secret, data: data)
+  end
+
+  def delete_secret(secret)
+    api.delete_secret(org: org, secret: secret)
+  end
+
   # identities
 
   def fetch_identity(identity, a_gvc = gvc)
@@ -448,6 +460,10 @@ class Controlplane # rubocop:disable Metrics/ClassLength
 
   def fetch_policy(policy)
     api.fetch_policy(org: org, policy: policy)
+  end
+
+  def delete_policy(policy)
+    api.delete_policy(org: org, policy: policy)
   end
 
   def bind_identity_to_policy(identity_link, policy)